Inproved Field Identification
Changed the way fields were addressed, and then added a foreach loop to
assign query string values. This could be more flexible, as the script
can look in a form for all possible Lasso values and then assign them to
a value in the curlData array. For example, naming a field email would
trigger the addition of that field as Emals[Primary] in the query
string.